made several fixes and updates

+++ qsc_food/models.py
@@ -53,11 +53,11 @@ class Cuisine(models.Model):
 class Food(models.Model):
     cuisine = models.ForeignKey(Cuisine, on_delete=models.CASCADE)
     name = models.CharField(max_length=128)
-    calories = models.IntegerField()
-    protein = models.IntegerField()
-    fibre = models.IntegerField()
-    sugar = models.IntegerField()
-    fat = models.IntegerField()
+    calories = models.FloatField()
+    protein = models.FloatField()
+    fibre = models.FloatField()
+    sugar = models.FloatField()
+    fat = models.FloatField()
 
 class Plate(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
